+++ src/generator-factory.js
@@ -1,13 +1,6 @@
 const glslTransforms = require('./glsl/composable-glsl-functions.js')
 const GlslSource = require('./glsl-source.js')
 
-//const renderpassFunctions = require('./glsl/renderpass-functions.js')
-
-Array.prototype.fast = function (speed) {
-  this.speed = speed
-  return this
-}
-
 class GeneratorFactory {
   constructor ({
       defaultUniforms,